// src/types/BlockchainQueryMethod.ts
export enum BlockchainQueryMethod {
  getbestblockhash = 'getbestblockhash',
  getblock = 'getblock',
  getblockchaininfo = 'getblockchaininfo',
  getblockcount = 'getblockcount',
  getblockhash = 'getblockhash',
  getblockheader = 'getblockheader',
  getblockstats = 'getblockstats',
  getchaintips = 'getchaintips',
  getchaintxstats = 'getchaintxstats',
  getdifficulty = 'getdifficulty',
  getmempoolancestors = 'getmempoolancestors',
  getmempooldescendants = 'getmempooldescendants',
  getmempoolentry = 'getmempoolentry',
  getmempoolinfo = 'getmempoolinfo',
  getrawmempool = 'getrawmempool',
  gettxout = 'gettxout',
  gettxoutproof = 'gettxoutproof',
  gettxoutsetinfo = 'gettxoutsetinfo',
  verifychain = 'verifychain',
  verifytxoutproof = 'verifytxoutproof'
}